requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-dirtyfields==1.9.0
django-filter==22.1
django-phonenumber-field==7.0.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.0.0
Markdown==3.4.1
phonenumbers==8.13.3
Pillow==9.4.0
PyJWT==2.6.0
whitenoise==6.3.0
drf-spectacular==0.25.1
python-magic-bin==0.4.14
django-autoslug==1.9.8
Faker==16.4.0
django-cleanup==6.0.0